Ensure aliases are resolved for HostingServiceMgr
This change makes it so that the HostingServiceMgr will be provided a
correct url for the remote repository.

This is because the remote URL stored in the configuration file might
be subject to user configuration, specifically this change is for
aliases.

SEE: `man -P 'less +/--get-url +n' git-ls-remote` for context on how we
resolve this URL.